requirements.txt
yfinance
pandas==1.5.3
pandas_ta
pandas_datareader
statsmodel
matplotlib
scikit-learn
PyPortfolioOpt
jupyterlab
arch
Flask 
flask_cors
alpaca-py
requests
python-dotenv
streamlit
numpy
datetime
scikit-learn
pandas_datareader
statsmodels

